Changes to how temp directories are created

+++ src/ModelingSystem/ModelingSystemGAMS.cpp
@@ -292,28 +292,25 @@ void ModelingSystemGAMS::createModelFromProblemFile(const std::string& filename)
     assert(modelingObject == nullptr);
     assert(modelingEnvironment == nullptr);
 
-    /* create temporary directory, mkdtemp is preferred over tmpnam */
-#if _DEFAULT_SOURCE || _BSD_SOURCE || _POSIX_C_SOURCE >= 200809L
-    strcpy(buffer, "loadgmsXXXXXX");
-    tmpdirname = mkdtemp(buffer);
-    if(tmpdirname.length() == 0)
-    {
+    if(env->settings->getSetting<bool>("Debug.Enable", "Output"))
+        tmpdirname = Utilities::createTemporaryDirectory(
+            "SHOT_GAMS_", env->settings->getSetting<std::string>("Debug.Path", "Output"));
+    else
+        tmpdirname = Utilities::createTemporaryDirectory("SHOT_GAMS_");
+
+    if(tmpdirname == "")
         throw std::logic_error("Could not create temporary directory.");
-    }
-#else
-    tmpdirname = std::tmpnam(nullptr);
-    fs::filesystem::create_directory(tmpdirname);
-    fs::filesystem::permissions(tmpdirname, fs::filesystem::perms::all);
-#endif
 
     createdtmpdir = true;
 
     /* create empty convertd options file */
     std::ofstream convertdopt((fs::filesystem::path(tmpdirname) / "convertd.opt").string(), std::ios::out);
+
     if(!convertdopt.good())
     {
         throw std::logic_error("Could not create convertd options file.");
     }
+
     convertdopt << " " << std::endl;
     convertdopt.close();
 
@@ -630,8 +627,8 @@ void ModelingSystemGAMS::clearGAMSObjects()
         auditLicensing = nullptr;
     }
 
-    /* remove temporary directory content (should have only files) and directory itself) */
-    if(createdtmpdir)
+    /* remove temporary directory contents if not in debug mode (should have only files) and directory itself) */
+    if(createdtmpdir && !env->settings->getSetting<bool>("Debug.Enable", "Output"))
     {
         fs::filesystem::remove_all(tmpdirname);
         createdtmpdir = false;
